upB flip adjustments
